+++ frontend/src/components/common/card/DefaultCardStyle.tsx
