+++ frontend/src/containers/ModelTypePage/ModelBannerContainer.tsx
@@ -1,7 +1,7 @@
 import { styled } from 'styled-components';
 import { BodyKrRegular4, BodyKrRegular5, HeadingKrRegular2 } from '../../styles/typefaces';
 import CenterWrapper from '../../components/layout/CenterWrapper';
-import Banner from '../../components/banner/Banner';
+import Banner from '../../components/common/banner/Banner';
 import HmgTag from '../../components/hmgTag/HmgTag';
 import PriceStaticBar from '../../components/priceStaticBar/PriceStaticBar';
