+++ frontend/src/containers/ModelTypePage/ModelSelectContainer.tsx
@@ -10,10 +10,10 @@ import {
   HeadingKrMedium7,
   HeadingKrRegular1,
 } from '../../styles/typefaces';
-import DefaultCardStyle from '../../components/card/DefaultCardStyle';
+import DefaultCardStyle from '../../components/common/card/DefaultCardStyle';
 import { useState } from 'react';
 import HmgTag from '../../components/hmgTag/HmgTag';
-import { CheckIcon } from '../../components/icons/Icons';
+import { CheckIcon } from '../../components/common/icons/Icons';
 import PriceSummary from '../../components/summary/PriceSummary';
 
 export default function ModelTypelSelectContainer() {
