+++ frontend/src/containers/OptionPage/OptionBannerContainer.tsx
@@ -8,11 +8,11 @@ import {
   HeadingKrRegular2,
 } from '../../styles/typefaces';
 import CenterWrapper from '../../components/layout/CenterWrapper';
-import Banner from '../../components/banner/Banner';
+import Banner from '../../components/common/banner/Banner';
 import HmgTag from '../../components/hmgTag/HmgTag';
 import PriceStaticBar from '../../components/priceStaticBar/PriceStaticBar';
 import { HTMLAttributes, useState } from 'react';
-import { ArrowLeft, ArrowRight } from '../../components/icons/Icons';
+import { ArrowLeft, ArrowRight } from '../../components/common/icons/Icons';
 
 interface ISubOptionTab extends HTMLAttributes<HTMLDivElement> {
   options: string[];
